+++ cmd/app/blockchain.go
@@ -94,7 +94,7 @@ func tronReclaimResourceCMD() *cli.Command {
 			st := store.New(psql)
 
 			// init system service
-			systemSvc := system.New(st, version, commitHash)
+			systemSvc := system.New(l, st, version, commitHash)
 			pID, err := systemSvc.ProcessingID(ctx)
 			if err != nil {
 				return fmt.Errorf("processing ID: %w", err)